+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,0 +1,34 @@
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { CartItem } from '../../models/cart-item.model';
+import { HighlightDirective } from 'src/app/shared/directives/highlight.directive';
+
+@Component({
+  selector: 'app-cart-list',
+  templateUrl: './cart-list.component.html',
+  styleUrls: ['./cart-list.component.scss']
+})
+export class CartListComponent {
+
+  @Input() cartItems!: CartItem[];
+  @Input() totalSum!: number;
+  @Input() totalQuantity!: number;
+  @Output() QuantityDecrease = new EventEmitter<CartItem>();
+  @Output() QuantityIncrease = new EventEmitter<CartItem>();
+  @Output() DeleteItem = new EventEmitter<CartItem>();
+
+  onQuantityIncrease(item: CartItem) {
+    this.QuantityIncrease.emit(item);
+  }
+
+  onQuantityDecrease(item: CartItem) {
+    this.QuantityDecrease.emit(item);
+  }
+
+  onDeleteItem(item: CartItem) {
+    this.DeleteItem.emit(item);
+  }
+
+  trackByItems(index: number, item: CartItem) {
+    return item.product.name;
+  }
+}