+++ src/app/cart/models/cart-item.model.ts
@@ -1,0 +1,8 @@
+import { Product } from "src/app/shared/models/Product";
+
+export class CartItem {
+    constructor(
+        public product: Product,
+        public quantity: number
+    ) { };
+}