+++ src/app/products/products.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ProductListComponent } from './product-list/product-list.component';
+import { FirstComponent } from './first/first.component';
+import { SharedModule } from '../shared/shared.module';
+
+
+
+@NgModule({
+  declarations: [
+    ProductListComponent,
+    FirstComponent
+  ],
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
+  exports: [
+    ProductListComponent
+  ]
+})
+export class ProductsModule { }