+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,21 @@
+import { Directive, ElementRef, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+
+  constructor(private el: ElementRef) { }
+
+  @HostBinding('style.backgroundColor') backgroundColor: string = '';
+
+  @HostListener('mouseenter') onMouseEnter() {
+    this.backgroundColor = '#FFFFE0 !Important';
+    this.el.nativeElement.style.cursor = 'pointer';
+  }
+
+  @HostListener('mouseleave') onMouseLeave() {
+    this.backgroundColor = '';
+    this.el.nativeElement.style.cursor = 'auto';
+  }
+}