+++ src/app/shared/models/Product.ts
@@ -3,11 +3,13 @@ export class Product {
     name: string;
     description: string;
     price: number;
+    isAvailable: boolean;
 
-    constructor(id: number, name: string, description: string, price: number) {
+    constructor(id: number, name: string, description: string, price: number, isAvailable: boolean) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.price = price;
+        this.isAvailable = isAvailable
     }
 }